document-process-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.10
botocore==1.38.10
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.4
dotenv==0.9.9
einops==0.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.27.1
lmdb==1.6.2
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.12.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
onnx==1.17.0
onnxoptimizer==0.3.13
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddle2onnx==2.0.1
paddlefsl==1.1.0
paddlenlp==2.6.1
paddlepaddle==3.0.0
pandas==2.2.3
parso==0.8.4
pdf2docx==0.5.8
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
polygraphy==0.49.20
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.25.5
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
rarfile==4.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
seqeval==1.2.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
